+++ src/utilities/graph/nodeMergeAddIds.ts
@@ -1,5 +1,6 @@
 import { VocabularyNode } from "../../types/types";
 import { isNil } from "../general/isNil";
+import { isArray } from "../general/isArray";
 
 /**
  * @ignore
@@ -10,8 +11,18 @@ export function nodeMergeAddIds(
   property: string
 ) {
   if (!isNil(newNode[property])) {
-    for (const arrayElement of newNode[property]) {
-      if (!oldNode[property].includes(arrayElement)) {
+    let newValues = newNode[property]
+    if(!isArray(newValues)){
+      // make sure the new values are used as array in the following algorithm
+      newValues = [newValues]
+    }
+    for (const arrayElement of newValues) {
+      // make sure the old value is changed to an array if needed
+      if(!isArray(oldNode[property])){
+        if(oldNode[property] !== arrayElement){
+          oldNode[property] = [oldNode[property], arrayElement]
+        }
+      } else if (!oldNode[property].includes(arrayElement)) {
         // add new entry
         oldNode[property].push(arrayElement);
       }